rename frost.go and move config and instance params to separate file
